+++ scrapli_netconf/channel/channel.py
@@ -70,7 +70,7 @@ class NetconfChannel(Channel, NetconfChannelBase):
         self._send_return()
         self._post_send_client_capabilities(capabilities_version=capabilities_version)
 
-    def send_input_netconf(self, channel_input: str) -> str:
+    def send_input_netconf(self, channel_input: str) -> bytes:
         """
         Send inputs to netconf server
 
@@ -78,7 +78,7 @@ class NetconfChannel(Channel, NetconfChannelBase):
             channel_input: string of the base xml message to send to netconf server
 
         Returns:
-            str: string result of message sent to netconf server
+            bytes: bytes result of message sent to netconf server
 
         Raises:
             N/A