+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 CHANGELOG
 =========
 
+# 2020.XX.XX
+- First real release??? :)
+
 # 2020.04.19
 - Initial pypi release... very beta still